requirements.txt
python-telegram-bot==21.3
python-docx
pandas
openpyxl
httpx
